requirements.txt
PyQt5>=5.15.0
watchdog>=3.0.0
Pillow>=9.0.0
